+++ src/main/kotlin/de/lrprojects/monaserver/controller/MemberController.kt
@@ -8,17 +8,17 @@ import org.springframework.http.ResponseEntity
 
 class MemberController : MembersApi {
     override fun apiGroupsGroupIdMembersPost(
-        groupId: Int?,
+        groupId: Long?,
         apiGroupsGroupIdMembersPostRequest: ApiGroupsGroupIdMembersPostRequest?
     ): ResponseEntity<Group> {
         return super.apiGroupsGroupIdMembersPost(groupId, apiGroupsGroupIdMembersPostRequest)
     }
 
-    override fun deleteMemberFromGroup(groupId: Int?): ResponseEntity<Void> {
+    override fun deleteMemberFromGroup(groupId: Long?): ResponseEntity<Void> {
         return super.deleteMemberFromGroup(groupId)
     }
 
-    override fun getGroupMembers(groupId: Int?): ResponseEntity<MutableList<Member>> {
+    override fun getGroupMembers(groupId: Long?): ResponseEntity<MutableList<Member>> {
         return super.getGroupMembers(groupId)
     }
